Added back old detection to fix queueing when accounts hit lvl 10 or 50
When an account reach lvl 10 or 50, the AFK Bot will not default to queuing for standard, until a human stops the bot and fixes the "New Playlist" or "New Event" button in the lobby. To fix this you simply go to the new gamemode, then queue for it for a few seconds, then stop queue and it's fixed.